Итоговое домашнее задание/game-pc-store-fullstack/backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.25,<2.0
psycopg2-binary>=2.9.1
pydantic>=1.8.0,<2.0
python-jose>=3.3.0
passlib>=1.7.4
alembic>=1.7.0
pytest>=6.2.5
httpx>=0.19.0
email-validator>=2.0.0